Moved roles + streaks logic into another file
